Optimize AERONET retrieval and consolidate tests
- Implemented chunked retrieval in `AERONETReader` (defaulting to 8 chunks) to prevent timeouts and NASA GSFC throttling.
- Consolidated all AERONET tests into `tests/test_aeronet.py` and removed redundant `tests/test_aeronet_aero.py`.
- Verified chunking logic with new unit tests.
- Maintained all previous improvements (protocol compliance, Dask support, robust retries, Pandas 3.0 compatibility).
